feat : 타이머 뷰 가즈아
